Split combined logs page into separate Call Logs and SMS Logs pages
Replaces the single /logs page with /logs/calls and /logs/sms, each linked
from the nav bar, so call and SMS history can be viewed independently.
