+++ frontend/src/router/index.js
@@ -1,11 +1,13 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Dashboard from '../views/Dashboard.vue'
-import Logs from '../views/Logs.vue'
+import CallLogs from '../views/CallLogs.vue'
+import SmsLogs from '../views/SmsLogs.vue'
 import Notifications from '../views/Notifications.vue'
 
 const routes = [
   { path: '/', component: Dashboard },
-  { path: '/logs', component: Logs },
+  { path: '/logs/calls', component: CallLogs },
+  { path: '/logs/sms', component: SmsLogs },
   { path: '/notifications', component: Notifications },
 ]
