+++ frontend/src/views/CallLogs.vue
@@ -3,7 +3,6 @@ import { ref, onMounted } from 'vue'
 import { api } from '../api'
 
 const callLogs = ref([])
-const smsLogs  = ref([])
 const loading  = ref(true)
 const error    = ref('')
 
@@ -14,20 +13,9 @@ const callHeaders = [
   { title: 'Timestamp', key: 'timestamp' },
 ]
 
-const smsHeaders = [
-  { title: 'ID',        key: 'id',        width: '80px' },
-  { title: 'User',      key: 'user' },
-  { title: 'Number',    key: 'number' },
-  { title: 'Message',   key: 'message' },
-  { title: 'Type',      key: 'sms_type' },
-  { title: 'Timestamp', key: 'timestamp' },
-]
-
 onMounted(async () => {
   try {
-    const [callData, smsData] = await Promise.all([api.getCallLogs(), api.getSmsLogs()])
-    callLogs.value = callData.data
-    smsLogs.value  = smsData.data
+    callLogs.value = (await api.getCallLogs()).data
   } catch (e) {
     error.value = e.message
   } finally {
@@ -40,7 +28,7 @@ onMounted(async () => {
   <v-container fluid class="pa-2 pa-sm-6">
     <v-alert v-if="error" type="error" class="mb-4">{{ error }}</v-alert>
 
-    <v-card class="mb-6">
+    <v-card>
       <v-card-title>
         <v-icon start>mdi-phone</v-icon>
         Call Logs
@@ -48,21 +36,6 @@ onMounted(async () => {
       <v-data-table
         :headers="callHeaders"
         :items="callLogs"
-        :loading="loading"
-        :items-per-page="20"
-        :sort-by="[{ key: 'timestamp', order: 'desc' }]"
-        :mobile-breakpoint="600"
-      />
-    </v-card>
-
-    <v-card>
-      <v-card-title>
-        <v-icon start>mdi-message-text</v-icon>
-        SMS Logs
-      </v-card-title>
-      <v-data-table
-        :headers="smsHeaders"
-        :items="smsLogs"
         :loading="loading"
         :items-per-page="20"
         :sort-by="[{ key: 'timestamp', order: 'desc' }]"
